Removed base wizard.
We don't need the base wizard. Just a base-controller and base error.

Using the base wizard in the way it was implemented meant we couldnt actually create a new instance of controller we were just patching the origional from hmpo-form-wizard.

This fixes up that issue by passing the controller into the wizard as a setting.