Eliminate reliance on Bash
Replaces the brace expansion in the create:public script in package.json.

Brace expansions are a Bash extension. npm runs its scripts through sh not bash.
On most modern GNU/Linux distributions sh is symlinked to bash which is why the
brace expansion will work in most cases. However, this is not always the case;
for example Ubuntu symlinks sh to dash which does not support bash completion.
This means that 'npm run create:public' fails and with it 'npm install'.