Prevent silent failure + Code clean-up
Currently sending the case-worker e-mail can fail silently, for example
when sending to an invalid address. This commit makes the app log an
error and spit it out to the user to help alert us to the problem.

Also, as the attachments are shared between both customer and
case-worker e-mails this code can been given its own 'var' to reduce
duplication.