[BRP-122] Correct nominee DOB in email
Displays the correct nominee DOB in the email for the "someone-else"
journey, by preventing the data being over written with junk.

"Wrong nominee DOB in "Someone else" journey e-mail

When selecting "Nominate someone else to collect my BRP" on step one of
the "someone else" journey the DOB you enter for the nominee will be
replaced, in both the customer and case-worker e-mails, with today's
date.

This is probably important as presumably this information is is used to
verify the nominee's date of birth when they come to collect they
BRP."