Remove potential empty heading
If there isn't a nominated person, the confirmation page shows an empty h1,
which could potentially be confusing for people relying on screen readers.

On top of that, it could be good to re-evaluate the heading structure of the
page again. It seems to jump between h1 and h3, and could be useful to
re-arrange them as h1-h2-h3 or just use h2's but add h3 classes to them
for visual purposes.